another day another over engineering a solution to a problem i invented
another day another over engineering a solution to a problem i invented

basically it will switch to the highest non doh/dol job if your on a doh/dol job. this is just a char selection screen QOL feature and completely pointless.
mainly in my case it will not leave characters on fishers on cleanup days. after running it once with the setting on you can just turn it off for subsequent runs as you won't have any chars in need of the job switch.

TESTED and working btw